requirements.txt
# Django and REST Framework
django==4.2.8
djangorestframework==3.14.0
django-model-utils==4.3.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# File Storage and Images
django-storages==1.14.2
boto3==1.33.7
Pillow==10.1.0
django-imagekit==5.0.0

# Environment and Configuration
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1

# Celery and Redis
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Rate Limiting
django-ratelimit==4.1.0

# Import/Export
django-import-export==3.3.5
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Monitoring and Logging
sentry-sdk==1.39.1
django-prometheus==2.3.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0